requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
Brotli==1.1.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.5
cssselect2==0.8.0
distro==1.9.0
ecdsa==0.19.1
exceptiongroup==1.3.0
faiss-cpu==1.11.0
fastapi==0.116.1
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.5.1
google-auth==2.40.3
google-genai==1.26.0
greenlet==3.2.3
groq==0.30.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
jinja2==3.1.6
joblib==1.5.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
ollama==0.5.1
packaging==25.0
passlib==1.7.4
pillow==11.3.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydyf==0.11.0
pygments==2.19.2
PyMySQL==1.1.1
pyphen==0.17.2
python-dotenv==1.1.1
python-jose==3.5.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.42
starlette==0.47.2
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.2
torch==2.2.2
tqdm==4.67.1
transformers==4.53.2
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0
weasyprint==66.0
webencodings==0.5.1
websockets==15.0.1
zopfli==0.2.3.post1
